Update 68.xml
In second paragraph of first response, "while it should be simplified" functions sort of as the opposite of a premise - I left it out bc it doesn't seem to fit either tag category, but it does have a function in the argument as agreement with the initial point.

Also, was I supposed to tag (rhetorical) questions? I just left those out.